Split homepage into multiple reusable components. Add a bunch of snapshots.
