+++ client/src/main/scripts/components/homepage/HomeSection.spec.tsx
@@ -1,0 +1,8 @@
+import * as React from "react";
+import * as renderer from "react-test-renderer";
+import HomeSection from "./HomeSection";
+
+test("HomeSection", () => {
+  const tree = renderer.create(<HomeSection />).toJSON();
+  expect(tree).toMatchSnapshot();
+});
