+++ client/src/main/scripts/components/homepage/HowToUseThis.spec.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+import * as renderer from "react-test-renderer";
+import HowToUseThis from "./HowToUseThis";
+
+jest.mock("./HomeSection", () => ({
+  __esModule: true,
+  default: "HomeSection",
+}));
+
+test("HowToUseThis", () => {
+  const tree = renderer.create(<HowToUseThis />).toJSON();
+  expect(tree).toMatchSnapshot();
+});
