--- client/src/main/scripts/components/homepage/infoHeader.ts
@@ -1,16 +1,0 @@
-export interface Tab {
-  tabTitle: string;
-  component?: React.Component;
-}
-
-export const INFO_HEADER_TABS: Tab[] = [
-  {
-    tabTitle: "How to use this",
-  },
-  {
-    tabTitle: "What is this",
-  },
-  {
-    tabTitle: "Get involved",
-  },
-];
